Updates to HTTP-AJAX code refactored for better structure
